+++ src/dipdup/exceptions.py
@@ -83,6 +83,23 @@ class DatasourceError(Error):
         """
 
 
+@dataclass(repr=False, kw_only=True)
+class AbiNotAvailableError(Error):
+    """ABI for the contract is not available"""
+
+    address: str
+    typename: str
+
+    def _help(self) -> str:
+        return f"""
+            ABI for the contract `{self.address}` is not available.
+
+            Check the contract address and datasource configuration. If contract is not verified, place ABI manually to `abi/{self.typename}/abi.json` and run `dipdup init`.
+
+            See https://dipdup.io/docs/datasources/abi_etherscan
+        """
+
+
 @dataclass(repr=False)
 class InvalidRequestError(Error):
     """API returned an unexpected response"""